Minor text, documentation, and link update.
I modified index.html so that, when JS is enabled, the RSS feed link is
set to use the RSS path defined in releases.js. I added a note to the
Readme documentation that this link needs to be updated when the site is
moved to a new server (to work smoothly when JS is disabled or
unavailable). I also made a minor text change to the text that is
displayed on generate-rss.html when JS is unavailable.