Appearance changes.
I modified the Show All Changes button so that it would appear blue when
active and added hover text. I also modified the Date filter so that
none of the dates are selected when it is inactive. I fixed a minor bug
in IE in which the expand button on the Date filter select is sized
correctly. Aside from this I made numerous small CSS revisions to font
weight, color, etc.